Improvements in Bash scripts

+++ timesheet.sh
@@ -1,15 +1,22 @@
 #!/usr/bin/env bash
 set -eou pipefail
-cd `dirname $0`
+cd "`dirname "$0"`"
 
-begin=${1:-09}
-end=${2:-13}
-files=$(eval echo data/2021-{$begin..$end}.yaml)
+source ./common.sh
 
-for f in $files
+initial_date=${1:-$(date +%F)}
+final_date=${2:-$initial_date}
+begin=$(get-week-number-for-date $initial_date)
+end=$(get-week-number-for-date $final_date)
+
+#inital_year=$(get-first-year-in-data-files)
+#final_year=$(get-last-year-in-data-files)
+# TODO: fix this for ...
+for f in $(eval echo data/2021-{$begin..$end}.yaml)
 do
+	[ -f $f ] || continue
 	ln -sf $f data.yaml
-	echo "Timesheet for file \"$f\":"
+	echo "Timesheet for week $(get-week-from-file-name $f):"
 	./timesheet.py
 	echo
 done
